Added Instrumentation
I used GCC's code generation facility to add automatic instrumentation to my code.
I'm thinking about splitting it off into a separate project because there really
isn't anything that is project-specific. I'm pretty excited about it, I think the
next thing to work on is a Python script to interpret the data and display it nicely.