Pre-Cleanup
Added a new class, GLRenderer, which holds a GLWindow and a matrix stack. This
might go away in the next commit, I'm starting to have second thoughts.

Turned off function profiling and instrumentation. I'm convinced that I'll have
access to that information when I need it, but until there's actually something that
needs profiling, the feature is just a waste of memory.