Adding Backbone
This is an intermediate commit. Added a new class, Dispatch, which acts as coordinator and router
for the various subsystems.

Also of note is a Python source file, rconsole.py. This module makes it easy to provide an
interactive Python console for a remote host. I forsee this being a super useful feature once I
embed a Python interpreter, which will probably be soon.