Added UnitTest++ to Project
The headers and library are statically linked to the project, the whole mess is contained as a
redistributable package in lib/UnitTest++. As new unit tests are written, they should be placed in the
tests folder, and they will be run as the 'tests' target in make. The debug target also depends upon
the tests target, so debug builds will run the unit tests.