Added basic texturing
The new shader in the 'res' folder is a flat color texture shader. It does not use any kind of
lighting, it just applies a color texture. The Mesh class now supports normal lighting and application
of a single texture. It desperately needs some TLC, it really ought to be like 3 or 4 different
classes and needs to work in a more general style.