Added Uniform class
Uniform management is now handled by defining structs which inherit BlockDefinition, populating the
fields of the struct, and passing it to a Uniform object. The BlockDefinition knows the name of the
uniform block it represents, how to query subsystems for the data it needs, and how to push that data
to a buffer on the GPU. The Uniform class is really just a thin wrapper which does housekeeping for
the GPU buffer.

Notably, the user of the Uniform class is still expected to manage allocation of buffer binding
points, similar to the Texture and Sampler classes.