Remote Console works
The branch says 'python', implying that everything python belongs here now, but
I feel like getting the remote console working is enough for this branch. The
important pieces are in place, now comes the real work of creating the
scripting API. That task will itself most likely be done in several branches,
so as to avoid a massive code dump and merge later in development.

As to what fixed the bug from last time.... I have no idea. Maybe I ran
cmake again? Perhaps a required shared object library was unavailable during
a system update? All I know is that I rebuilt the project and it worked, with
no indication of any sort of error.

Changes in this commit:
-------------------------------------------------------------------------------

    - Instantiate RemoteConsole in the Dispatch class, update during the main
        loop. Works like a charm