Added Unit Tests to Core
I finished up the test suite for Vectors, and decided not to do Quaternions.
Even though writing tests takes relatively little time, my development time is
limited, so I'm only going to write tests for code which satisfies some or all
of the following criteria:

    - Suspected (or known) to be buggy or error-prone
    - Has potential to create silent errors
    - Tests can be written without a major refactoring
    - Complex bit of code which does significant computation.

Notably, these criteria rule out most of the code in the rendering subsystem,
as a lot of that code boils down to being a wrapper around the OpenGL API.
I can see code in the physics system and the scripting interface being tested,
but testing is no longer a priority.