Minor Edits to Build System
The CMakeLists.txt was updated a bit to do a better job of autodetecting the
presence (or lack thereof) of required libraries. The idea is for CMake to spit
out a useful error rather than the build fail because a required header is
missing. The name of the header for UnitTest++ was changed to reflect the
library's location after being built from the AUR.