Sync with 'dev'
Commiting local changes in anticipation of a sync with 'dev'. Forking the edits
branch off of py-api, while convenient, proved to be a mistake. I can't continue
to build the project without those changes and some new ones that popped up.

Apparently updating GLM was a BAD IDEA. The updates deprecated degrees as
arguments, forcing a ton of warnings to be vomitted all over my terminal.